Refactor common views and fix formset Add button issue
- Refactored common views for better code organization.
- Fixed issue where clicking the Add button in the formset model was triggering the user menu instead.
