+++ src/common/views/__init__.py
@@ -1,0 +1,3 @@
+from .crud_view import CRUDView
+from .table_view import CustomTableView
+from .common_views import CUSTOM_FIELD_VIEWS
