+++ src/templates/account/logout.html
@@ -2,7 +2,7 @@
 
 {% block content %}
 <!-- Logout Section -->
-<section class="w-full h-full flex justify-center items-center absolute bg-gray-700 bg-opacity-50 top-0 left-0 right-0 bottom-0 z-50">
+<section class="w-full h-full flex justify-center items-center absolute bg-gray-700 bg-opacity-50 top-0 left-0 right-0 bottom-0 z-40">
   <!-- Container -->
   <div class="md:w-2/4 h-fit bg-white rounded-lg shadow md:mt-0 sm:max-w-md xl:p-0">
     <!-- Close Button -->
